Make comparison match intent
The comparison is intended to check if FirstOrDefault did not find an
element. An equality comparison should be made against default(double),
because that's the value returned by FirstOrDefault when no element is
found.

This commit possibly also fixes an unconfirmed bug where the comparison
fails on some architectures, specifically the one used on iPhone 4 and on
generation 5 iPod touch.