Axis should always honor absolute minimum /maximum as well as minimum/maximum range properties (#1819)
* Added failing tests for OxyPlot issue 1812

An axis with absolute limits and MinimumRange set will not always honor
these after a call to Reset().  Two new tests added which demonstrate
the isse, one without data and the other with data partially
out-of-range.

* Check for valid MinimumRange

The exception message for the check that MinimumRange is not larger than the difference of absolute maximum and absolute minimum was wrong. Added also a check for MinimumRange not being larger than MaximumRange

* Fix OxyPlot issue 1812

Axis constraints like absolute minimum, absolute maximum, minimum range and maximum range should now work even with no data or data which is outside the absolute limits. The validity checks have been moved to the start of the method so that there is no partial setting of properties.

* Changelog and Contributors updated